Remove redundant onBlur event handler
Can't really remember why I needed this in the first place (so it might come back) but this is causing weird issues with state update (e.g. it will update the title of the form to the initial load value when losing focus (presumably because it captures that value only once), and a sidebar click has no effect while focused).