Select published version in publish tab
In the publish tab, we will show the currently published version of the prompt if there is one rather than the currently selected one. If you unpublish, you will be able to publish the currently selected one instead (if it has runs). The alternative would be to have some kind of republish or update button in addition to the publish checkbox.