Decide which flavour to use based on header
We can add more decision logic later on.

"the decision on which one to use may be a binary one (e.g. toggle the entire project from dev to prod) or it may be based on an x-environment header in the client request (e.g. you have a separate iOS build for employees) or it may be a staged roll-out where you route 10% of the requests to prod and the rest to dev, or it may depend on a unique identifier in the request like a user ID"