Give partial runs an ID
This is mainly so we can scroll to them by setting the active run ID without too much special casing. Feels a bit hacky but these low range indices should never coincide with actual datastore IDs so it should be safe. However we still need to make sure we don't allow commenting for partial runs so we disable the selection handler when the inputs field is missing (previously the identifier would be undefined but now it's always there because we need to be able to select the element when scrolling). Hence also renamed selectionIdentifier back to identifier because it is used for both selection and auto scrolling.