Don't require user project to have a user id
Now that we introduce workspaces, we no longer need to use a special ID as a sentinel for the "Prompts" project (aka user project). Instead we will have a "Drafts" workspace for each user with the user ID. To avoid clashes and complicating the access table, we want to change the ID of those old user projects, but that means we need another way to distinguish the Prompts project in the old UI (only until we have the new UI).