Add back button in project sidebar
Will add it to top bar later but full width top bar looks a bit silly with the old tabs design so will need to clean that up first. The challenge is implementing the back button for a shared project because we may not have access to that workspace so we need to detect that case.