Remove unused move prompt functionality for now
I think we may want to implement Copy to Project rather than Move to Project anyway, because otherwise there could be weird implications on endpoints and chains the prompt is used in etc (although we still need to deal with those when deleting the prompt). If we do decide to reimplement it, we also need to think about which projects we allow to move to (only within workspace vs all workspaces).