Factor out EndpointsView
This was hard but an important step. We can now use EndpointsView for the publish tab in both prompt and chain tab view, with lazy loading of the active prompt (so it doesn't presume this to be preloaded). Next step will be to move it to the project top level (where we don't have fully loaded prompts).