Simplify post-duplication logic
Selecting the new item after duplication doesn't feel that important (not sure we even want it, especially if you start thinking about generalising it to copy to other project/workspace). And if we don't need it, we can get rid of quite a few callbacks that no longer need to be passed in.