Save individual inputs on chain entity
Realised yesterday that this will make things easier e.g. if we want to show a proper patch editor without first having to load all the prompts to get the inputs. We already store the inputs on the chain entity to allow showing the curl command in the endpoints view, so we may as well store them properly per chain item rather than aggregated as a separate column.