Remove prompt selector from publish settings pane
This was a bit confusing anyway (e.g. you have a Prompt / Chain label and then a Prompt label right below). We are changing it so that you pick the prompt or chain when adding the endpoint and don't change it afterwards (instead just add another endpoint if you want to publish another prompt or chain).