Avoid redundant rich text updates
This (mainly) fixes the issue where the cursor would jump to the end of the html content when removing a newline before a non-empty line, but the root cause is that we shouldn't modify the raw HTML content unless we have to, i.e. as long as the normalised, sanitised HTML stays the same, we don't really care what other garbage is added internally. This fix is similar to previous ones but works at a level deeper (where the first sanitisation happens).