Strip formatting from HTML control
On the one hand, we don't want to retain any formatting (e.g. style/font attributes) except for some very specific instances (i.e. input variables). On the other hand, we don't really want to touch the raw HTML because that may cause the cursor to jump to the end. The core of the problem is that the ContentEditable component we use here seems to insert some redundant formatting in specific cases like removing a newline. Stripping that formatting causes the cursor to jump (which we don't want) but not stripping any formatting isn't an option either. So for now we will have to live with this occasional cursor jump (and maybe consider switching to a different library later).