Make sure ref is loaded when we focus
Because this component is JS heavy, we load it dynamically using Suspense, but that means we can't assume the ref will already be loaded when the parent component is mounted. So we add a special callback as a ref that will be called when the node we need is mounted (inspired by https://legacy.reactjs.org/docs/hooks-faq.html#how-can-i-measure-a-dom-node).