Don't rely on old project to load new prompt
The toActivePrompt logic was broken in a subtle way because we only loaded the new versions and then built up the active prompt from those versions and the project data, but the prompt in the project data might not have been refreshed yet (e.g. lastVersionID might be wrong at this point after saving).