Fix popup when starting with empty versions
We should make sure the container ref is set when we start with empty versions too. At some point we may also want to make hooks like useContainer less fragile (so it works when the ref is not immediately loaded too). This issue is similar to what we had with suspense-loaded containers but the problem is (still) that functional refs aren't compatible with the useResizeObserver hook that we use.