Don't save empty prompt as new version
We already disable the Run button for an empty prompt, but that means any saving of an empty prompt happens implicitly (i.e. not when hitting Run) and that can just be skipped as it's not useful at all to go from a non-empty prompt to a completely empty prompt version.