Get rid of editing state in code node editor
Now that saving is an explicit action, we can just update the items whenever we make a change in the code editor (previously we certainly didn't want to avoid save on each keystroke). This is a lot simpler. However we can still have an edited prompt to deal with when running the chain, so we still need the saving logic to allow passing in items different to the current items.