Use timestamp to designate intermediate results
Currently we rely on the indices to determine which partial runs were intermediate results (and hence can stay on screen after completing the run). This is a bit fragile especially if we get additional steps (indices) because of interrupts. It's better to just let the backend mark the runs that should be persisted by adding/removing the timestamp.