Don't track prompt input scroll height
This was a bit fragile anyway because it would only increase and never decrease the needed scroll height. Let's just give it its minimum height and allow you to make it higher if you want more room for the prompt. This also means we no longer need to constrain the prompt height because the minimum is an actual minimum.