Workaround for line-clamp not working in Safari
Close #59

This isn't perfect cause you don't see the ellipsis in Safari but it's a lot better. Also slightly changes the behaviour in the sense that clicking a small cell in the row will now also expand the height if other cells in that row need it, but that doesn't feel too bad.