Don't include same system prompt multiple times
In the case of running the same (chat) prompt multiple times, we don't want to include the system prompt over and over again. However you could imagine a more complex use case where you do include context of previous prompts but still change or tweak the system prompt in a later prompt in the chain, and in that case we should include it. Easiest solution is to just check whether we have already included this same system prompt before and not add it again in that case.