Overwrite last version even when branching
Basically, we will no longer create multiple versions without runs for the same user and always overwrite the last (only) version without runs even when we go back to another source version and make edits from there. This requires to send the source ("previous") version to the api in addition to the one we're going to overwrite ("current"). The latter is for efficiency so we don't need to search it on the server when we already know it on the client.