Move user versions without runs to the back
We can't really assume that a version without runs will be the most recent one (even though we now always overwrite the single last version without runs for that user), because there could be other users that have run a version after you saved yours. So we need to make sure that we expose the version without runs from the current user at the end of the list.