Move some logic into ChainNodeBox component
On the one hand it's tempting for the node component not to have to be aware of its own index and the other nodes. On the other hand, passing in all those helpers feels a bit excessive. On top of that, there are instance where one node actually does need to know about other nodes, e.g. to show unbound inputs in the inputs node (of course you could pass that in too but it starts to feel cumbersome).