Fix edge case in active version tracking
We previously changed this because e.g. if we push a new URL on deletion, the initial active item and version may change. But to track the initial state properly here, we should compare the active item ID and not the active version ID. Otherwise saving a version on selecting a new one breaks (it jumps back to the saved version after save rather than staying on the selected one as we want).