Add migration to fill in missing chain versions
These might occur due to bug fixed in previous commit, and would be fixed on loading and saving the chain again but we can't count on that exactly (e.g. in theory you could create an endpoint without doing that). Maybe not worth it but might avoid issues down the road.