Don't strip isolated breaks in divs
Usually the <br /> tags appear at the end of a <div>, in which case they are redundant. But if you shift-enter, you get an isolated <br /> tag which we should not strip but just map to a newline after converting the divs to lines. These will be converted to regular divs next roundtrip so they are not canonical (see added tests).