Pretty-print JSON returned from endpoint
This should not make any functional difference but feels a bit nicer and is more consistent with how it works with streaming enabled (e.g. OpenAI will stream formatted JSON). We do need to set the header to ensure we are still returning JSON when using res.send instead of res.json.