Fix salting of continuation IDs
This was a bit careless as we need to ensure a BigInt xor operation is used here (and then we don't need to worry about ending up with unsigned either). There should be no risk of losing precision when casting back to Number (even if XOR did overflow which it should not?), because at that point we should have arrived at the original ID again (and when it's a new continuation ID, we cast the BigInt directly to String  on line 132).