Clear query params after editing endpoint
If we created an endpoint from a version, the parent and version ID will be in the query. We want to remove these in case you either cancel or finish creating that new endpoint, so that a reload (e.g. to see new analytics) doesn't trigger the endpoint creation sheet again.