Don't refresh prompt on rename
This fixes a bug where renaming a prompt would revert pending changes to a prompt. In that case we don't really need to refresh the prompt but just the project to update the sidebar. When renaming a chain we do need to refresh the chain itself because we show the name in the editor tab (however for a chain you are less likely to end up with pending changes because we save more frequently). This also feels a bit cleaner overall, being explicit about what we want to refresh on each project item action.