Don't include users in both access list & pending
In theory you could invite someone to a project first and then to the workspace, in which case they could end up with access to the project but not the workspace (in the vice versa case, they would be able to access the project even if they still have pending invite for the project). In both these cases, we should show them in the list with people who have access to the project and not in the pending invitations list.