Fix for rebuilding active item on refresh project
We need to be careful here that we don't capture the old active item but instead update the current one after refreshing the project (e.g. when we are first refreshing the active item and then the project in a single callback that was constructed before calling the api and hence captures the state at that point).