Fix edge case in cursor position restoration
If we assign a count zero to empty nodes (think empty lines), we can't really distinguish between multiple empty lines (i.e. we don't know if we were on the first or second empty line before the cursor jumped to the end). We can count 1 instead, but then we need to be careful to also compensate for this while restoring (as we can't actually jump to position 1 on an empty line). Feels a bit hacky but can't immediately think of something more elegant.