Refresh active version on refresh project
We already refresh the active prompt or chain when we refresh the project (cause some of the prompt or chain properties come from the project), but we should also refresh the version, because e.g. the comments on the versions may have changed as they come from the project. This fixes the issue where a new comment on the content of a run wouldn't immediately be rendered as highlighted in the UI.