Add IsSiblingNode helper
The goal is to represent a tree with branches as a linear list of nodes (or put differently, to add branching to our existing linear list of chain items). This helper is the core that everything is based on, i.e. it's possible to decide locally whether a node is at the same level as the previous one or occurs on the next level. This is not super obvious but it's not too hard to prove that this is a necessary and sufficient condition: a node is a sibling IFF its branch number is greater than the branch number of the previous node added to the number of extra branches introduced by the previous node.