Don't send project invite emails to waitlist users
We already register the invitation to workspace or project in case they ever get access, but we should not send an invitation email in this case. Note that we're taking a little shortcut by checking if the user has access to their own Drafts workspace (which will only be created once the user has access to the product).