Ignore pending invitations when deleting project
Having unaccepted pending invitations on a project or workspace should not prevent you from deleting that project or workspace. Access keys have been generalised to be either default or pending, so we can't just look at the number of access keys here. Getting the project or workspace users is slightly more expensive but deleting a project or workspace is not a very frequent action.