Allow to delete project shared with others
I first thought that the issue was on the client side (showing the "Delete" action rather than "Leave" in case you have shared a project with others), but it's actually the semantics on the backend that are wrong (cause they should not mimic those of workspaces). Leaving a workspace can make sense even if you created it, but leaving a project you created does not really make sense (there is no access key to revoke in this case, and what workspace would the project continue to live in if it first lived in your Drafts?). Instead we will allow to delete a project even if you have saved it with others, but not if someone else shared it with you (in that case the available action is still to leave it rather than delete it). This is now different from workspaces, where you can only leave a workspace once you have shared it with others (and only the last user can delete it).