Keep runs sorted as they go from partial to final
There are actually 3 different states rather than just 2 (partial vs persisted): ongoing, completed, and persisted. The ongoing state was already sorted by input (test row), but the others weren't. To keep the sorting in the persisted state, we need to save the runs in order. But to also keep the sorting in the intermediated completed-but-not-persisted state, we need to wait with sending a timestamp until the end.