Don't prevent mapping same input twice in chain
This is clearly incorrect when e.g. you use the same prompt in two separate branches and you want to map one of its inputs in a preceding node on each branch (i.e. both nodes should be able to map to the same output). But even without branches this feels a bit too restrictive, e.g. if you want to use the same input in consecutive nodes and overwrite the previous output.