Generate single empty input set if all rows empty
This fixes an issue where we enable the run button with empty test data if the only variables are dynamic ones (functions), but we don't actually run the prompt because the list of PromptInputs is empty. The fact that this works when running a prompt without variables is due to rowCount being -Infinity rather than 0 in that case, which feels very fragile and most likely wasn't deliberate. Instead we will just make this consistent and return a single empty PromptInputs in both those cases (so technically you could now also run a prompt with static variables and empty test data if only the button was enabled which is fine).