Mark additional older versions for a user as run
Usually there should be at most one version per user that hasn't been run yet, but there could be older versions from before we changed that behaviour. It is better to mark these as run (and have them show up in version history and be exposed to other users) than to leave them in this ghost state because they might pop up at unexpected times with old labels and comments when we overwrite them (and we allow to overwrite them because they have never been run).