Output continuation key when streaming too
Up until now, the way you would implement chat would be to provide your own unique continuation key. This has the nice (or confusing, depending on how you look at) side effect that we don't need to tell you the continuation key, so we can just stream the output which will be very easy to consume because you don't need to parse it. However, we now want to change it so chat mode is an explicit attribute of the prompt. You don't need to provide your own unique continuation key in this case (which is good), but you probably still want to be able to use streaming for chat, and we also need to tell you what the continuation key is (at least once after you sent the first message). So we will just stream it as an extra line which is already formatted as a curl header (and easy to parse otherwise). For now I've decided not to include this when the continuation key doesn't change (which it won't for consecutive messages in the same conversation). The advantage is that this way, you can *still* provide your own unique continuation key upfront and never need to parse it from the response (but I imagine this will be more of a hidden/undocumented feature rather than the recommended way of implementing chat). Note that this also means you can now turn on streaming when using function callbacks although this may be less important (less likely to happen in interactive environment). Previously, this would not work because the continuation key you needed to respond was only included when streaming was disabled.