Strip quotes around variables in code blocks
If you write

return "{{variable}}"

Is it unlikely that you want the result to be equal to "variable" and more likely that you want the variable to be expanded and returned as a string (though the former is according to spec if you think about what happens in this case exactly). If you really wanted the result to be "variable", you would very likely not have included the braces in the first place. Hence we will assume that quotes around variables in code blocks are unintentional/redundant and strip them. Added some tests which also illustrate that you don't really lose flexibility by doing this.