Trigger continuation after running chat prompt
At this point, you can run a chat endpoint without providing your own unique continuation key, but instead by toggling the prompt used in the endpoint as being in chat mode. You can't do this from the UI yet because we don't expose a continuation key on the runs yet (and don't provide a way to continue either).