Persist continuationID from response in run entity
We may want to extend this later so that we also save the continuation ID when it was passed in but no longer exists on the response (like for endpoints), so we can still link the last run in a continuation to its predecessors, although then we may need a way to distinguish this last one because we don't want to show the input box in that case. For now we won't worry about that because with the use case of chat the continuation can't just stop at some point anyway.