Don't expire continuation cache for UI runs
In the case of endpoints, we expire the continuation cache after 1 hour (maybe we should make it a day, but the idea is that we don't want to store tons of old conversations if we have all the individual logs as well). But in the case of runs, you might actually come across an old conversation in the UI and decide to continue it (and also the number of these conversation should be at least an order of magnitude smaller than the number of endpoint requests for most real world products, so the load of storing these will be much less, and they will get deleted when the version/prompt/chain/project gets deleted).