Use text-bison rather than text-bison@001
This is more similar to what we do for OpenAI and Anthropic, where we don't lock to a specific version but use the latest version of Claude Instant or GPT 4.

See https://cloud.google.com/vertex-ai/docs/generative-ai/model-reference/text

"To use the latest model version, specify the model name without a version number, for example text-bison.

To use a stable model version, specify the model version number, for example text-bison@001. Each stable version is available for six months after the release date of the subsequent stable version."