Don't reset isChat flag on running migration again
Shot myself in the foot with this cause I thought I found a save version bug but it was due to running migration again to update the bison model and setting isChat to false again by doing that. Shouldn't happen in prod as we will only run the migration once per version but better safe than sorry. Also adding some logging to detect multiple pending versions.