Timestamp previous partial runs when clearing last
When we clear the last partial runs (because they get replaced by the properly persisted runs in the prompt responses), we also want to ensure that previous partial runs appear before those persisted runs rather than after. So here we assign them an artificial timestamp that respects their inherent order just to accomplish that. Note that this only really matters when testing functions in the UI, because that is the only case where a single prompt can generate multiple steps per input.