Allow sending continuation reply as plain text
Feels a bit redundant having to specify the "message" key in the body and having to use application/json as content type if you are just replying with a plain message. This shorthand will only work for chat (not functions), and you still need to specify content type text/plain for it to work in curl. We could also relax this for functions (where we currently require the body key to be equal to the function name) but in all of this there is probably a trade-off between being more implicit for convenience vs being more explicit to allow easier debugging when things don't go as expected (we may even already be crossing that line here).