Make 16k GPT3.5 point to latest model
According to documentation, gpt-3.5-turbo-16k "Will point to gpt-3.5-turbo-1106 starting Dec 11, 2023". At that point the regular gpt-3.5-turbo will ALSO point to that same model (so there will be little point in having a separate 16k model and we will likely merge them together). For now, the approach of only remapping the 16k model to the new one seems to offer most flexibility, as you can now either use the very latest model or keep using the current GPT 3.5 Turbo for another month, and the price for the 16k will be lower than before.