Extend default cache expiration time
The main goal here is that we don't end up with a huge amount of cached conversations from endpoint requests that will never be accessed again but that we can't really remove because the endpoint stays active. However, it doesn't really matter whether we clean them up after an hour or a day, so we may as well extend the default expiration in the rare case you are testing something and actually want to continue an endpoint conversation after more than an hour.