Persist whether run can be continued or not
See comments in 2b99fe9afd942e55b041f5b3752383564a33283e and e2d0e51d81033a905b889f0b704cbeb1f4a016f5. We want to be able to link all runs who share the same continuationID, but we also need to be able to distinguish between the ones that can use this continuationID to continue vs the ones who cannot because they come at the end of a continuation. This wasn't an issue with chat mode but it is for function interrupts.