Only show reply box if we can in fact continue
This fixes a minor issue where in theory you could already reply to a partial run (which was itself a continuation) while it was still running. But more importantly this avoids that you can reply to any run (partial or not) which is itself (the last) part of a continuation but doesn't have the option to continue the continuation any further.