Make continuation more consistent with first run
While we don't save the prompt or chain when running a continuation (and we don't run multiple inputs either), we still want the rest of the behaviour to be consistent, e.g. save test data before running and update the active index in the chain while running (we may want to update this latter bit but that needs to happen in general so either way we want it to be consistent).