Add prompt config context
This is important to ensure that we always reflect the updated config (i.e. we can't just rely on the initial default prompt config and assume we will refresh when needed). On the other hand, we only need to set up this context when we are on a project item.