Remove support for gpt-4-32k
This isn't used in production at all (it is only exposed if your OpenAI account has access to it and you need to go through special hoops to get that) and it now seems obsolete because you can just use gpt-4-turbo (preview for now) instead which should be better in all regards (price, speed, context window). Keeping the gated model logic in for now but it won't actually add any models.