Switch to test mode on selecting input or output
Been going back and forth on this one, but when you select input or output, it most likely means you want to inspect inputs or outputs and hence switch to test mode (and having to do that manually as an extra step feels annoying). The trade-off is that you may need to switch back to Edit mode manually if you proceed to focus on another node and were in that mode before, but that feels less bad and less common.