Filter out unsupported prompts when changing model
If you switch from a model which supports system prompts to a model which doesn't, the system prompt should be cleared. This wasn't really an issue before because there was no way to see those and they wouldn't get used anyway but with the new version history you would be able to see them.