Ignore undefined entries when comparing chains
This is tricky. When you JSON.stringify an object with undefined values, it will strip them out first, but if you JSON.stringify the array of entries of that object, it will include them with null values. This was causing a tricky bug where the client would think that the version had been changed (not ignoring the stripped-by-setting-to-undefined activePrompt and version attributes) and then the server would recognise that the version hasn't changed so it would overwrite the previous one, resulting in potentially only the author being updated for that old version.