Fix issue with redundant prompt fetching
The active item cache wasn't deduplicating the prompt IDs so we're seeing way to many getPrompt api calls when there are a lot of chain versions. We don't even really need to load those prompts, it just helps to show the version IDs in the diff which is not critical (so let's limit it to 10).