Filter out null providers if alternative available
When we load all providers for a given scope or a given user, we include the ones with null keys (these previously had an api key that got then removed, and we want to retain their metadata in case they are reinstated). But when we need to load credentials for a single provider based on multiple scopes, we need to make sure a provider with valid key is returned if available.