Don't show global provider cost
We're going to start tracking this in a more granular, scoped way, so keeping a single cost per provider feels obsolete (and might not reflect what you expect because your actual usage may be spread across your user and different projects. Was never part of the design to show this in Settings anyway (and it's not critical in admin user dashboard either).