Don't check cache in case of continuation
We don't cache responses if there is a continuation (in input or output), so checking the cache is a bit pointless in that case. For now we will just assume that caching and interrupts are incompatible (otherwise we would need to consider the persisted state in addition to the inputs which could get complex in the general case).