Persist userID of project creator
Not needing this is kind of elegant, but not having it has bitten us a few times now (e.g. when adding admin dashboards), and now again, cause we need to use basic heuristics to assign an owner for each project. Let's avoid this kind of issue in the future by simply keeping track of who created each project (even though the creator could end up assigning a different owner and leave the project altogether eventually).