Fix edge case in global popup position
Usually, you would specify either top or bottom position, and either left or right position. But the exception is when the popup is fixed width. In that case there may be a conflict between adjusted right, unadjusted left, and child width. In this case the left and width will win out and the right will be ignored. But that means we should adjust the left as well even though the right is specified. Should probably write some tests for this logic at some point.