Require ownership for project settings
This is securing the backend (functionality should already be hidden on client but that doesn't prevent an attacker from calling an endpoint directly). It is convenient that we are already using ensureScopeAccess to guard half of these call sites (and reassuring that the other call sites had a TODO). Note that it is more efficient to check for project ownership than for project access because we don't need to consider the encompassing workspace.