Fix shared projects query
The logic used to be that whenever there was an access key for a project, it meant that you were invited to it and hence it was a shared project. But now there access keys for all projects you own as well (including all the ones you created yourself), so these should clearly be excluded from the list of shared projects. On the other hand, you might have been made owner of a project that was shared, so these need to stay included in the shared project list. So the right filter is not to look at project ownership but rather to exclude the projects you have implicit access to through the workspace.