Restrict move and delete project to owners
Now that we expose which projects are owned by the user, we can make this more accurate and secure. Basically, only project owners should be able to delete a project or move it to another workspace. If you don't own a project, you can leave it but only if the project was explicitly shared and not if you can simply access it through the workspace (otherwise you should leave workspace instead).