Fix saving chain on reverting to previous version
The callback that is passed to savePrompt is only executed when the prompt actually needs saving. So while it is fine to update the prompt cache only in that case, we can't rely on this to update our chain version because we may need to do that even if the prompt doesn't need saving, because you may have made changes to the prompt so that it now matches a previously saved version of that prompt.