Fix focus version on prompt/chain selection
We allow to specify a version when selecting or refreshing a prompt/chain, but this doesn't actually have any effect because the useInitialState logic will reset it again (because the initialActiveVersion changes when the activeItem changes). The main way this surfaced is by selecting a comment (it wouldn't actually focus on the right version but instead jump to the last one if it was in a different prompt/chain).