Highlight selected response
Right now it is only possible to focus on a response by selecting a comment, but in that case we will highlight the run in the run timeline similar to a version in the version timeline. We also allow selecting a different response in that case, which doesn't make much sense right now but it will in the case of run responses once you can inspect individual steps.