Fix streaming of interrupted chain in endpoint
When running a chain, we don't want to stream output of intermediate nodes but just the final result. However, when the chain is interrupted, we need to output the intermediate result (function call) even when streaming is on. But we can't stream that output in this case because we don't know upfront whether running a prompt will trigger an interrupt or not (in order to stream it, we would need to know that upfront). So the best we can do in this case is to output the result of the interrupted node without streaming it.