Fix edge conditions in intermediate run fetching
Took me a long time to get this right. Basically we need to avoid a race condition but also fetch with the parent ID rather than the active run ID (which could be a later continuation run, and we need the first parent ID to retrieve the initial intermediate responses as those are the only ones that won't have the continuation ID yet).