Remove function response from inputs after step
If we just completed a continuation step we should clear the function response that was consumed in this step just in case the same function call is used later in the chain. More specifically this will keep the InterruptOnce logic in the code block working when the same function call is used in consecutive code blocks. Another example would be where the same function call is used in consecutive prompts and we don't want to reuse the previous response if a response is missing from a next response (this is more of an edge case though).