Don't save chain without prompt version known
This fixes an edge case where, if you add a prompt to a chain that isn't in the cache yet, and then run the chain without making any additional changes (i.e. not hitting Save and Close on the prompt either), it would persist and then try to run the prompt config without a version in it which of course fails. A better solution might be to subscribe to the cache and save once the ID of the latest version of that prompt is known, but given you didn't click Save and Close, just postponing the save is probably good enough in practice (also evidenced by how long it took to find this bug).