Send timestamp on step completion
Now that we detect step completion (to be able to save intermediate runs), we can just send a timestamp at that point as well and no longer need to stream it after completing the entire chain (which would only set a timestamp on the last partial run for each input). This also simplifies the logic for consuming the streamed data of partial runs.