Don't assign timestamp to partial runs
We can now assign accurate timestamps to partial runs immediately after each step in the chain, but actually that means we can't avoid them from jumping around when ordered by timestamp if running a chain with multiple parallel inputs. On the other hand, we can now just keep them sorted without timestamps because we no longer interleave partial runs with persisted runs now that we have intermediate persisted runs (i.e. we clear the partial runs on completion). Note that not streaming the timestamps anymore means we won't have a timestamp for failed runs (as they never get persisted) but that should be fine.