Fix edge case in node connector UI
This fixes the visual connector when there are no chain items and you switch to test mode. If we just passed in isDisabled on the previous line, we could also just pass in isActive on the next line. But because we have another case which we treat as isDisabled (i.e. isActive and hasSpotOnNextRow), we should also adjust the value of isActive to consider the state of isDisabled (because there can't really be an active menu index when you are in test mode).