Fix another edge case in merge run logic
There were actually two bugs in this one line. We shouldn't update the run id unless previous run which was partial was actually used to represent the group. And when we don't update the run id, we should retain the id of the run that currently represents the group and not of the continuation we are comparing with.