Don't expose previous month cost
We reset the cost when we first check the budget on running a prompt in the new month. But if you look at your usage before running a prompt in the new month, it will still show the cost from last month. Alternatively, we could checkBudgetForScope before getTrustedBudgetForScope, but simply reusing the same logic in the toBudget method avoids the extra query (and potential save at this point).